backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-genai
assemblyai

